+++ pymeasure/instruments/rohdeschwarz/__init__.py
@@ -25,3 +25,4 @@
 from .fsseries import FSL, FSW
 from .hmp import HMP4040
 from .sfm import SFM
+from .smp04 import SMP04